"a lot of changes, nothing works XD"

+++ printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   printf.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: marvin <marvin@student.42.fr>              +#+  +:+       +#+        */
+/*   By: mseara <mseara@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/05 13:51:00 by mseara            #+#    #+#             */
-/*   Updated: 2022/10/12 12:41:27 by marvin           ###   ########.fr       */
+/*   Updated: 2022/10/16 10:20:21 by mseara           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,13 @@
 
 #include <stdarg.h>
 #include <unistd.h>
+//borrar next_line
+#include <stdio.h>
 #include "./libft/libft.h"
 
 int	ft_printf(char const *, ...);
 int	ft_collect_args(char const *str, va_list args);
 int	ft_putchar(int c);
 int	ft_putstr(char *str);
-int ft_puthex(int num, int len);
+int ft_puthex(unsigned long long num, int len);
 #endif
